Refactor: place assets' api-related code inside `api` subpackage
